# APIdoc/api.yaml
openapi: 3.0.3
info:
  title: Documentazione API progetto "SatisTrento"
  description: Documentazione API per il progetto "SatisTrento" all'interno del progetto "100 progetti per il comune di Trento" per il corso di Ingegneria del Software a.a. 2024/2025 corso di laurea in Informatica presso l'Università degli Studi di Trento.
  version: 0.0.1
servers:
  - url: http://localhost:3000/api/v1
    description: Development server for testing
components:
  schemas:
...
    User:
      description: Oggetto contenente le informazioni di un utente.
      type: object
      properties:
        self:
          type: string
          description: Url della risorsa.
        email:
          type: string
          description: Email dell'utente.
        nome:
          type: string
          description: Nome dell'utente.
        cognome:
          type: string
          description: Cognome dell'utente.
        ruolo:
          type: string
          enum:
            - "amministratore"
            - "analista"
            - "sondaggista"
            - "circoscrizione"
          description: Ruolo dell'utente.
        imageUrl: 
          type: string
          description: Url dell'immagine profilo dell'utente. (gravatar)
...
    Error:
      description: Oggetto contenente le informazioni di un errore.
      type: object
      properties:
        code:
          type: integer
          description: Codice dell'errore.
        message:
          type: string
          description: Messaggio di errore.
        details:
          type: string
          description: Dettagli dell'errore.
          nullable: true
    AddSondaggio:
      description: Oggetto contenente le informazioni per aggiungere un sondaggio.
      type: object
      properties:
        titolo:
          type: string
          description: Titolo del sondaggio.
        dataInizio:
          type: string
          description: Data di inizio del sondaggio.
          format: date
    SondaggioBase: 
      description: Oggetto contenente le informazioni di base di un sondaggio.
      type: object
      allOf:
        - $ref: "#/components/schemas/AddSondaggio"
      properties:
        _id:
          type: string
          description: Identificatore univoco del sondaggio.
        isAperto:
          type: boolean
          description: Indica se il sondaggio è aperto (in corso) o chiuso.
        statoApprovazione:
          type: string
          enum:
            - "approvato"
            - "in attesa"
            - "rifiutato"
          description: Stato di approvazione del sondaggio.
        sondaggista:
          "$ref": "#/components/schemas/User"
    ElencoSondaggi:
      description: Array di oggetti contenenti le informazioni di base di un sondaggio (visualizzazione sondaggista/admin).
      oneOf:
        - type: array
          items:
            "$ref": "#/components/schemas/SondaggioBase"
        - type: array
          items:
            "$ref": "#/components/schemas/Sondaggio"
    AddVoto:
      description: Oggetto contenente le informazioni per aggiungere un voto.
      type: object
      properties:
        eta:
          type: integer
          description: Età dell'utente che ha effettuato il voto.
          minimum: 0
        voto:
          type: integer
          description: Voto assegnato al quartiere.
          minimum: 1
          maximum: 5
        quartiere:
          type: integer
          description: Identificatore univoco del quartiere.
    Voto:
      description: Oggetto contenente le informazioni di un voto.
      type: object
      properties:
        _id:
          type: string
          description: Identificatore univoco del voto.
        dataOra:
          type: string
          description: Data e ora in cui è stato effettuato il voto.
          format: date-time
      allOf:
        - "$ref": "#/components/schemas/AddVoto"
    MediaVoti:
      description: Oggetto semplice contenente la media dei voti di un quartiere.
      type: object
      properties:
        media:
          type: number
          description: Media dei voti assegnati al quartiere.
          format: double
          minimum: 1
          maximum: 5
        quartiere:
          type: integer
          description: Identificatore univoco del quartiere.
    Sondaggio:
      description: Oggetto contenente le informazioni di un sondaggio.
      allOf:
        - "$ref": "#/components/schemas/SondaggioBase"
      type: object
      properties:
        voti:
          type: array
          items:
            "$ref": "#/components/schemas/Voto"
          description: Array di oggetti contenenti i voti assegnati al sondaggio.
        mediaVoti:
          type: array
          items:
            "$ref": "#/components/schemas/MediaVoti"
          description: Array di oggetti contenenti la media dei voti assegnati ai quartieri presenti nel sondaggio (quelli non presenti non vengono visualizzati).
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Token di autenticazione fornito all'utente per accedere alle risorse specifiche protette.
paths:
...
  '/sondaggi' :
    description: "Risorsa che restituisce la lista dei sondaggi dell'utente corrente, se amministratore allora restituisce tutti i sondaggi."
    get: 
      summary: "Ottieni lista dei sondaggi"
      description: "Restituisce la lista dei sondaggi dell'utente corrente. Se l'utente è amministratore, restituisce tutti i sondaggi."
      security:
        - BearerAuth: []
      parameters:
        - name: deepData
          in: query
          description: "Se impostato a true, restituisce le informazioni complete di ogni sondaggio."
          required: false
          schema:
            type: boolean
            default: false
          allowEmptyValue: true
      responses:
        200:
          description: "Successful response"
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/ElencoSondaggi"
        401:
          description: "Utente non autenticato."
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/Error"
        403:
          description: "Utente loggato ma non ha i permessi per visualizzare i sondaggi."
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/Error"
      tags:
        - Sondaggi
        - Sondaggista
    post: 
      summary: "Crea un nuovo sondaggio"
      description: "Crea un nuovo sondaggio. Solo per utenti con ruolo 'sondaggista'."
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/AddSondaggio"
      responses:
        201:
          description: "Sondaggio creato con successo."
          headers:
            Location:
              description: "Url della risorsa creata."
              schema:
                type: string
        400:
          description: "Dati del sondaggio non validi."
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/Error"
        401:
          description: "Utente non autenticato."
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/Error"
        403:
          description: "Utente loggato ma non ha i permessi per creare un sondaggio."
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/Error"
      tags:
        - Sondaggi
        - Sondaggista
  '/sondaggi/{id}' :
    description: "Risorsa che restituisce le informazioni di un sondaggio."
    parameters:
      - name: id
        in: path
        description: "Identificatore univoco del sondaggio."
        required: true
        schema:
          type: string
    get: 
      summary: "Ottieni informazioni di un sondaggio"
      description: "Restituisce le informazioni di un sondaggio, solo per utenti con ruolo 'sondaggista' o 'amministratore'."
      security:
        - BearerAuth: []
      responses:
        200:
          description: "Successful response"
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/Sondaggio"
        401:
          description: "Utente non autenticato."
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/Error"
        403:
          description: "Utente loggato ma non ha i permessi per visualizzare il sondaggio."
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/Error"
        404:
          description: "Sondaggio non trovato."
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/Error"
      tags:
        - Sondaggi
        - Sondaggista
    patch: 
      summary: "Modifica un sondaggio"
      description: "Modifica un sondaggio. Solo per utenti con ruolo 'sondaggista'."
      security:
        - BearerAuth: []
      
      requestBody:
        required: true
        content:
          application/json:
            schema:
              allOf:
                - "$ref": "#/components/schemas/AddSondaggio"
              type: object
              properties:
                isAperto:
                  type: boolean
                  description: "Indica se il sondaggio è aperto (in corso) o chiuso."
      responses:
        200:
          description: "Sondaggio modificato con successo."
          headers:
            Location:
              description: "Url della risorsa modificata."
              schema:
                type: string
        400:
          description: "Dati del sondaggio non validi."
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/Error"
        401:
          description: "Utente non autenticato."
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/Error"
        403:
          description: "Utente loggato ma non ha i permessi per modificare il sondaggio."
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/Error"
        404:
          description: "Sondaggio non trovato."
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/Error"
      tags:
        - Sondaggi
        - Sondaggista
    delete:
      summary: "Elimina un sondaggio"
      description: "Elimina un sondaggio. Solo l'utente che ha creato il sondaggio può eliminarlo o un amministratore, se il sondaggio è in attesa. Se il sondaggio è approvato o rifiutato, solo l'amministratore può eliminarlo."
      security:
        - BearerAuth: []
      responses:
        204:
          description: "Sondaggio eliminato con successo."
        401:
          description: "Utente non autenticato."
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/Error"
        403:
          description: "Utente loggato ma non ha i permessi per eliminare il sondaggio."
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/Error"
        404:
          description: "Sondaggio non trovato."
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/Error"
      tags:
        - Sondaggi
        - Sondaggista
...